+++ Backend/app/helpers/payments_helper.rb
@@ -1,0 +1,2 @@
+module PaymentsHelper
+end
