+++ Backend/app/views/carts/show_current.html.erb
@@ -79,7 +79,7 @@
         </strong>
       </div>
 
-      <%= button_to 'Continuar compra', start_checkout_checkouts_path, method: :post,  class: 'btn btn-primary w-100'%>
+      <%= button_to 'Continuar compra', orders_path, method: :post,  class: 'btn btn-primary w-100'%>
     </div>
   <% end %>
 </section>
