+++ Backend/test/controllers/payments_controller_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class PaymentsControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
